Fix migration file for filling isProjectVerified
